sort.c: jump pos on count, for msb and lsb
During shuffling we loop on the array but jump on next bucket when we realize current bucket is sorted.
In the worst case the entire array gets sorted when we are on the first position, but still have to traverse the rest of the array to see if it is sorted. This commit adds jumping support so we make at most 255 more accesses before we terminate